+++ lsm/config/models/__init__.py
@@ -42,6 +42,7 @@ from .modes import (
 )
 from .query import QueryConfig
 from .chats import ChatsConfig
+from .agents import AgentConfig, SandboxConfig
 from .lsm_config import LSMConfig
 
 __all__ = [
@@ -83,5 +84,7 @@ __all__ = [
     "RemoteProviderChainConfig",
     "QueryConfig",
     "ChatsConfig",
+    "AgentConfig",
+    "SandboxConfig",
     "LSMConfig",
 ]